refactor: unify parser extensions into a single webpackSyntax plugin
Merge lazy locations, import attributes and import phases into one
acorn extension in lib/javascript/syntax.js (one prototype level, one
place to incrementally take over parsing). Import phases are switched
per parse via the importPhases option derived from the deferImport/
sourceImport experiments, replacing the process-global
JavascriptParser.extend call in WebpackOptionsApply.
