+++ lib/WebpackOptionsApply.js
@@ -60,7 +60,6 @@ const URLPlugin = require("./dependencies/URLPlugin");
 const WorkerPlugin = require("./dependencies/WorkerPlugin");
 
 const JavascriptModulesPlugin = require("./javascript/JavascriptModulesPlugin");
-const JavascriptParser = require("./javascript/JavascriptParser");
 
 const JsonModulesPlugin = require("./json/JsonModulesPlugin");
 
@@ -527,20 +526,6 @@ class WebpackOptionsApply extends OptionsApply {
 
 			const httpOptions = options.experiments.buildHttp;
 			new HttpUriPlugin(httpOptions).apply(compiler);
-		}
-
-		if (
-			!(
-				/** @type {typeof JavascriptParser & { __importPhasesExtended?: true }} */
-				(JavascriptParser).__importPhasesExtended
-			) &&
-			(options.experiments.deferImport || options.experiments.sourceImport)
-		) {
-			const { importPhases } = require("./javascript/syntax");
-
-			JavascriptParser.extend(importPhases);
-			/** @type {typeof JavascriptParser & { __importPhasesExtended?: true }} */
-			(JavascriptParser).__importPhasesExtended = true;
 		}
 
 		new EntryOptionPlugin().apply(compiler);
