+++ lib/javascript/JavascriptModulesPlugin.js
@@ -455,19 +455,28 @@ class JavascriptModulesPlugin {
 								case JAVASCRIPT_MODULE_TYPE_AUTO: {
 									return new JavascriptParser("auto", {
 										parse: options.parse,
-										typescript: options.typescript
+										typescript: options.typescript,
+										importPhases: Boolean(
+											options.deferImport || options.sourceImport
+										)
 									});
 								}
 								case JAVASCRIPT_MODULE_TYPE_DYNAMIC: {
 									return new JavascriptParser("script", {
 										parse: options.parse,
-										typescript: options.typescript
+										typescript: options.typescript,
+										importPhases: Boolean(
+											options.deferImport || options.sourceImport
+										)
 									});
 								}
 								case JAVASCRIPT_MODULE_TYPE_ESM: {
 									return new JavascriptParser("module", {
 										parse: options.parse,
-										typescript: options.typescript
+										typescript: options.typescript,
+										importPhases: Boolean(
+											options.deferImport || options.sourceImport
+										)
 									});
 								}
 							}
@@ -1956,11 +1965,16 @@ class JavascriptModulesPlugin {
 			}
 			const code = /** @type {string} */ (moduleSource.source());
 
+			const { experiments } = compilation.options;
 			const { ast } = JavascriptParser._parse(
 				code,
 				{
 					sourceType: "auto",
-					ranges: true
+					ranges: true,
+					// generated code contains phase imports when the experiments are on
+					importPhases: Boolean(
+						experiments.deferImport || experiments.sourceImport
+					)
 				},
 				JavascriptParser._getModuleParseFunction(compilation, m)
 			);
@@ -2028,11 +2042,16 @@ class JavascriptModulesPlugin {
 		for (const [m, moduleSource] of inlinedModuleSources) {
 			const code = /** @type {string} */ (moduleSource.source());
 
+			const { experiments } = compilation.options;
 			const { ast } = JavascriptParser._parse(
 				code,
 				{
 					sourceType: "auto",
-					ranges: true
+					ranges: true,
+					// generated code contains phase imports when the experiments are on
+					importPhases: Boolean(
+						experiments.deferImport || experiments.sourceImport
+					)
 				},
 				JavascriptParser._getModuleParseFunction(compilation, m)
 			);
