+++ lib/javascript/JavascriptParser.js
@@ -6,7 +6,7 @@
 "use strict";
 
 const vm = require("vm");
-const { Parser: AcornParser, tokTypes } = require("acorn");
+const { Parser: AcornParser } = require("acorn");
 const { HookMap, SyncBailHook } = require("tapable");
 const NormalModule = require("../NormalModule");
 const Parser = require("../Parser");
@@ -18,7 +18,11 @@ const {
 	webpackCommentRegExp
 } = require("../util/magicComment");
 const BasicEvaluatedExpression = require("./BasicEvaluatedExpression");
-const { SourcePositions, lazyLocations } = require("./syntax");
+const {
+	LEGACY_ASSERT_ATTRIBUTES,
+	SourcePositions,
+	webpackSyntax
+} = require("./syntax");
 
 /** @typedef {import("acorn").Options} AcornOptions */
 /** @typedef {import("acorn").ecmaVersion} EcmaVersion */
@@ -157,74 +161,8 @@ const ALLOWED_MEMBER_TYPES_CALL_EXPRESSION = 0b01;
 const ALLOWED_MEMBER_TYPES_EXPRESSION = 0b10;
 const ALLOWED_MEMBER_TYPES_ALL = 0b11;
 
-const LEGACY_ASSERT_ATTRIBUTES = Symbol("assert");
-
-/** @type {(BaseParser: typeof AcornParser) => typeof AcornParser} */
-const importAssertions = (Parser) =>
-	class extends Parser {
-		/**
-		 * Parses with clause.
-		 * @this {InstanceType<AcornParser>}
-		 * @returns {ImportAttribute[]} import attributes
-		 */
-		parseWithClause() {
-			/** @type {ImportAttribute[]} */
-			const nodes = [];
-
-			const isAssertLegacy = this.value === "assert";
-
-			if (isAssertLegacy) {
-				if (!this.eat(tokTypes.name)) {
-					return nodes;
-				}
-			} else if (!this.eat(tokTypes._with)) {
-				return nodes;
-			}
-
-			this.expect(tokTypes.braceL);
-
-			/** @type {Record<string, boolean>} */
-			const attributeKeys = {};
-			let first = true;
-
-			while (!this.eat(tokTypes.braceR)) {
-				if (!first) {
-					this.expect(tokTypes.comma);
-					if (this.afterTrailingComma(tokTypes.braceR)) {
-						break;
-					}
-				} else {
-					first = false;
-				}
-
-				const attr =
-					/** @type {ImportAttribute} */
-					this.parseImportAttribute();
-				const keyName =
-					attr.key.type === "Identifier" ? attr.key.name : attr.key.value;
-
-				if (Object.prototype.hasOwnProperty.call(attributeKeys, keyName)) {
-					this.raiseRecoverable(
-						attr.key.start,
-						`Duplicate attribute key '${keyName}'`
-					);
-				}
-
-				attributeKeys[keyName] = true;
-				nodes.push(attr);
-			}
-
-			if (isAssertLegacy) {
-				/** @type {EXPECTED_ANY} */
-				(nodes)[LEGACY_ASSERT_ATTRIBUTES] = true;
-			}
-
-			return nodes;
-		}
-	};
-
 // Syntax: https://developer.mozilla.org/en/SpiderMonkey/Parser_API
-let parser = AcornParser.extend(importAssertions, lazyLocations);
+let parser = AcornParser.extend(webpackSyntax);
 
 /** @typedef {Record<string, string> & { _isLegacyAssert?: boolean }} ImportAttributes */
 
@@ -367,6 +305,7 @@ class VariableInfo {
  * @property {boolean=} allowHashBang
  * @property {boolean=} allowReturnOutsideFunction
  * @property {SourcePositions=} lazySourcePositions internal: serve loc/range lazily instead of allocating them during parsing
+ * @property {boolean=} importPhases enable parsing of the import phase proposals (import defer / import source)
  */
 
 /**
@@ -540,7 +479,7 @@ class JavascriptParser extends Parser {
 	/**
 	 * Creates an instance of JavascriptParser.
 	 * @param {"module" | "script" | "auto"=} sourceType default source type
-	 * @param {{ parse?: ParseFunction, typescript?: boolean }=} options parser options
+	 * @param {{ parse?: ParseFunction, typescript?: boolean, importPhases?: boolean }=} options parser options
 	 */
 	constructor(sourceType = "auto", options = {}) {
 		super();
@@ -4995,7 +4934,8 @@ class JavascriptParser extends Parser {
 					locations: true,
 					ranges: true,
 					comments: true,
-					semicolons: true
+					semicolons: true,
+					importPhases: this.options.importPhases === true
 				},
 				this.options.parse
 			));
@@ -5052,7 +4992,10 @@ class JavascriptParser extends Parser {
 	evaluate(source) {
 		const { ast } = JavascriptParser._parse(
 			`(${source})`,
-			{ sourceType: this.sourceType },
+			{
+				sourceType: this.sourceType,
+				importPhases: this.options.importPhases === true
+			},
 			this.options.parse
 		);
 		if (ast.body.length !== 1 || ast.body[0].type !== "ExpressionStatement") {
@@ -5853,7 +5796,7 @@ class JavascriptParser extends Parser {
 			const comments = [];
 
 			// When both loc and range are wanted (the main parse path), have the
-			// lazyLocations plugin serve them on demand instead of eagerly.
+			// parser plugin serve them on demand instead of eagerly.
 			const sourcePositions =
 				options.locations && options.ranges
 					? new SourcePositions(code)
