+++ lib/javascript/syntax.js
@@ -11,6 +11,8 @@ const { tokTypes } = require("acorn");
 /** @typedef {typeof import("acorn").Parser} AcornParser */
 /** @typedef {import("acorn").Position} AcornPosition */
 /** @typedef {import("acorn").Parser} AcornParserInstance */
+/** @typedef {import("acorn").TokenType} TokenType */
+/** @typedef {import("estree").ImportAttribute} ImportAttribute */
 /** @typedef {import("estree").SourceLocation} SourceLocation */
 /** @typedef {[number, number]} Range */
 
@@ -20,6 +22,12 @@ const LINE_BREAK_G = /\r\n?|\n|\u2028|\u2029/g;
 // Symbol-keyed so it stays out of for-in, Object.keys and JSON.stringify
 // over AST nodes.
 const kSourcePositions = Symbol("source positions");
+
+// Marks import attributes parsed from the legacy `assert {...}` syntax.
+const LEGACY_ASSERT_ATTRIBUTES = Symbol("assert");
+
+// acorn's BIND_LEXICAL scope binding type
+const BIND_LEXICAL = 2;
 
 /**
  * Offset → line/column mapping for one parsed file. The line-start table is
@@ -138,23 +146,57 @@ class LazyLocNode {
 }
 
 /**
- * Runs acorn with `locations`/`ranges` disabled and serves `loc`/`range`
- * through `LazyLocNode` instead. Activated per parse via the internal
- * `lazySourcePositions` option; without it the parser behaves like plain
- * acorn.
+ * Methods and state internal to acorn, absent from its public types.
+ * @typedef {AcornParserInstance & {
+ * type: TokenType,
+ * value: EXPECTED_ANY,
+ * start: number,
+ * startLoc?: AcornPosition,
+ * containsEsc: boolean,
+ * next: () => void,
+ * eat: (type: TokenType) => boolean,
+ * expect: (type: TokenType) => void,
+ * afterTrailingComma: (type: TokenType, notNext?: boolean) => boolean,
+ * unexpected: (pos?: number) => never,
+ * raise: (pos: number, message: string) => never,
+ * raiseRecoverable: (pos: number, message: string) => void,
+ * isContextual: (name: string) => boolean,
+ * parseIdent: (liberal?: boolean) => EXPECTED_ANY,
+ * checkLValSimple: (expr: EXPECTED_ANY, bindingType?: number) => void,
+ * startNode: () => EXPECTED_ANY,
+ * startNodeAt: (pos: number, loc?: AcornPosition) => EXPECTED_ANY,
+ * finishNode: (node: EXPECTED_ANY, type: string) => EXPECTED_ANY,
+ * parseImport: (node: EXPECTED_ANY) => EXPECTED_ANY,
+ * parseImportSpecifiers: () => EXPECTED_ANY[],
+ * parseImportAttribute: () => EXPECTED_ANY,
+ * parseExprImport: (forNew: boolean) => EXPECTED_ANY,
+ * parseImportMeta: (node: EXPECTED_ANY) => EXPECTED_ANY,
+ * parseDynamicImport: (node: EXPECTED_ANY) => EXPECTED_ANY,
+ * [kSourcePositions]?: SourcePositions,
+ * _importPhase: "defer" | "source" | null,
+ * _importPhasesEnabled: boolean,
+ * }} ParserWithInternals
+ */
+
+/**
+ * webpack's single extension over acorn — the one place parser behavior is
+ * customized and the staging ground for incrementally taking over parsing.
+ * Covers lazy `loc`/`range` (via the internal `lazySourcePositions` option),
+ * import attributes (`with { ... }` and legacy `assert { ... }`) and import
+ * phases (via the `importPhases` option). Unlike the former
+ * `acorn-import-phases` package, acorn's `!forNew` guard is kept, so
+ * `new import.defer(...)` is a SyntaxError just like `new import(...)`.
  * @type {(BaseParser: AcornParser) => AcornParser}
  */
-const lazyLocations = (Parser) => {
-	/** @typedef {{ startLoc?: AcornPosition, startNode: () => EXPECTED_ANY, startNodeAt: (pos: number, loc?: AcornPosition) => EXPECTED_ANY }} AcornParserInternals */
-
-	// node creation methods are internal to acorn and absent from its types
-	const base =
-		/** @type {AcornParserInternals} */
-		(/** @type {unknown} */ (Parser.prototype));
+const webpackSyntax = (Parser) => {
+	// internal methods are absent from acorn's types
+	const base = /** @type {ParserWithInternals} */ (
+		/** @type {unknown} */ (Parser.prototype)
+	);
 
 	return class extends Parser {
 		/**
-		 * @param {AcornOptions & { lazySourcePositions?: SourcePositions }} options options
+		 * @param {AcornOptions & { lazySourcePositions?: SourcePositions, importPhases?: boolean }} options options
 		 * @param {string} input source code
 		 * @param {number=} startPos start position
 		 */
@@ -165,10 +207,14 @@ const lazyLocations = (Parser) => {
 			}
 			super(options, input, startPos);
 			this[kSourcePositions] = sourcePositions;
+			this._importPhase = null;
+			this._importPhasesEnabled = options.importPhases === true;
 		}
 
+		// ----- lazy loc/range -----
+
 		/**
-		 * @this {AcornParserInstance & { [kSourcePositions]?: SourcePositions, start: number }}
+		 * @this {ParserWithInternals}
 		 * @returns {EXPECTED_ANY} new node
 		 */
 		startNode() {
@@ -177,7 +223,7 @@ const lazyLocations = (Parser) => {
 		}
 
 		/**
-		 * @this {AcornParserInstance & { [kSourcePositions]?: SourcePositions }}
+		 * @this {ParserWithInternals}
 		 * @param {number} pos start offset
 		 * @param {AcornPosition=} loc start position when acorn tracks locations
 		 * @returns {EXPECTED_ANY} new node
@@ -190,7 +236,7 @@ const lazyLocations = (Parser) => {
 		/**
 		 * Mirror of acorn's `copyNode`, which bypasses `startNodeAt` via
 		 * `new Node(...)` and would otherwise produce non-lazy nodes.
-		 * @this {AcornParserInstance & AcornParserInternals}
+		 * @this {ParserWithInternals}
 		 * @param {EXPECTED_ANY} node node to copy
 		 * @returns {EXPECTED_ANY} copied node
 		 */
@@ -199,54 +245,70 @@ const lazyLocations = (Parser) => {
 			for (const prop in node) newNode[prop] = node[prop];
 			return newNode;
 		}
-	};
-};
 
-// acorn's BIND_LEXICAL scope binding type
-const BIND_LEXICAL = 2;
+		// ----- import attributes (`with { ... }` / legacy `assert { ... }`) -----
 
-/**
- * Methods and state internal to acorn, absent from its public types.
- * @typedef {AcornParserInstance & {
- * type: import("acorn").TokenType,
- * value: EXPECTED_ANY,
- * start: number,
- * startLoc?: AcornPosition,
- * containsEsc: boolean,
- * next: () => void,
- * eat: (type: import("acorn").TokenType) => boolean,
- * unexpected: (pos?: number) => never,
- * raise: (pos: number, message: string) => never,
- * raiseRecoverable: (pos: number, message: string) => void,
- * isContextual: (name: string) => boolean,
- * parseIdent: (liberal?: boolean) => EXPECTED_ANY,
- * checkLValSimple: (expr: EXPECTED_ANY, bindingType?: number) => void,
- * startNodeAt: (pos: number, loc?: AcornPosition) => EXPECTED_ANY,
- * finishNode: (node: EXPECTED_ANY, type: string) => EXPECTED_ANY,
- * parseImport: (node: EXPECTED_ANY) => EXPECTED_ANY,
- * parseImportSpecifiers: () => EXPECTED_ANY[],
- * parseExprImport: (forNew: boolean) => EXPECTED_ANY,
- * parseImportMeta: (node: EXPECTED_ANY) => EXPECTED_ANY,
- * parseDynamicImport: (node: EXPECTED_ANY) => EXPECTED_ANY,
- * _importPhase: "defer" | "source" | null,
- * }} ParserWithInternals
- */
+		/**
+		 * @this {ParserWithInternals}
+		 * @returns {ImportAttribute[]} import attributes
+		 */
+		parseWithClause() {
+			/** @type {ImportAttribute[]} */
+			const nodes = [];
 
-/**
- * Supports the import phase proposals: static `import defer * as ns from ...`
- * / `import source x from ...` set `phase` on the ImportDeclaration, dynamic
- * `import.defer(...)` / `import.source(...)` yield an ImportExpression with
- * `phase`. Replaces the `acorn-import-phases` package; unlike it, acorn's
- * `!forNew` guard is kept, so `new import.defer(...)` is a SyntaxError just
- * like `new import(...)`.
- * @type {(BaseParser: AcornParser) => AcornParser}
- */
-const importPhases = (Parser) => {
-	const base = /** @type {ParserWithInternals} */ (
-		/** @type {unknown} */ (Parser.prototype)
-	);
+			const isAssertLegacy = this.value === "assert";
 
-	return class extends Parser {
+			if (isAssertLegacy) {
+				if (!this.eat(tokTypes.name)) {
+					return nodes;
+				}
+			} else if (!this.eat(tokTypes._with)) {
+				return nodes;
+			}
+
+			this.expect(tokTypes.braceL);
+
+			/** @type {Record<string, boolean>} */
+			const attributeKeys = {};
+			let first = true;
+
+			while (!this.eat(tokTypes.braceR)) {
+				if (!first) {
+					this.expect(tokTypes.comma);
+					if (this.afterTrailingComma(tokTypes.braceR)) {
+						break;
+					}
+				} else {
+					first = false;
+				}
+
+				const attr =
+					/** @type {ImportAttribute} */
+					this.parseImportAttribute();
+				const keyName =
+					attr.key.type === "Identifier" ? attr.key.name : attr.key.value;
+
+				if (Object.prototype.hasOwnProperty.call(attributeKeys, keyName)) {
+					this.raiseRecoverable(
+						attr.key.start,
+						`Duplicate attribute key '${keyName}'`
+					);
+				}
+
+				attributeKeys[keyName] = true;
+				nodes.push(attr);
+			}
+
+			if (isAssertLegacy) {
+				/** @type {EXPECTED_ANY} */
+				(nodes)[LEGACY_ASSERT_ATTRIBUTES] = true;
+			}
+
+			return nodes;
+		}
+
+		// ----- import phases (`import defer/source`, `import.defer/source()`) -----
+
 		/**
 		 * @this {ParserWithInternals}
 		 * @param {EXPECTED_ANY} node import declaration node
@@ -264,6 +326,10 @@ const importPhases = (Parser) => {
 		 * @returns {EXPECTED_ANY[]} import specifiers
 		 */
 		parseImportSpecifiers() {
+			if (!this._importPhasesEnabled) {
+				return base.parseImportSpecifiers.call(this);
+			}
+
 			const phase = this.isContextual("defer")
 				? "defer"
 				: this.isContextual("source")
@@ -335,6 +401,7 @@ const importPhases = (Parser) => {
 			const node = base.parseExprImport.call(this, forNew);
 
 			if (
+				this._importPhasesEnabled &&
 				node.type === "MetaProperty" &&
 				(node.property.name === "defer" || node.property.name === "source")
 			) {
@@ -364,6 +431,10 @@ const importPhases = (Parser) => {
 		 * @returns {EXPECTED_ANY} MetaProperty node
 		 */
 		parseImportMeta(node) {
+			if (!this._importPhasesEnabled) {
+				return base.parseImportMeta.call(this, node);
+			}
+
 			this.next();
 
 			const containsEsc = this.containsEsc;
@@ -398,6 +469,6 @@ const importPhases = (Parser) => {
 	};
 };
 
+module.exports.LEGACY_ASSERT_ATTRIBUTES = LEGACY_ASSERT_ATTRIBUTES;
 module.exports.SourcePositions = SourcePositions;
-module.exports.importPhases = importPhases;
-module.exports.lazyLocations = lazyLocations;
+module.exports.webpackSyntax = webpackSyntax;
