+++ lib/optimize/ConcatenatedModule.js
@@ -2434,11 +2434,17 @@ ${defineGetters}`
 				let ast;
 
 				try {
+					const { experiments } = this.compilation.options;
+
 					({ ast } = JavascriptParser._parse(
 						code,
 						{
 							sourceType: "module",
-							ranges: true
+							ranges: true,
+							// generated code contains phase imports when the experiments are on
+							importPhases: Boolean(
+								experiments.deferImport || experiments.sourceImport
+							)
 						},
 						JavascriptParser._getModuleParseFunction(this.compilation, m)
 					));
